extract scrolling logic into a separate component
