+++ kolibri/core/assets/src/views/AppBar.vue
@@ -194,7 +194,6 @@
   }
 
   .wrapper {
-    z-index: 4;
     background-color: $core-action-normal;
   }
